Added a missing semicolon and changed file name.
