--- LCF.c
@@ -1,18 +1,0 @@
-#include <stdio.h>
-int main() {
-    int n1, n2, max;
-    printf("Enter two positive integers: ");
-    scanf("%d %d", &n1, &n2);
-
-    // maximum number between n1 and n2 is stored in min
-    max = (n1 > n2) ? n1 : n2;
-
-    while (1) {
-        if (max % n1 == 0 && max % n2 == 0) {
-            printf("The LCM of %d and %d is %d.", n1, n2, max);
-            break;
-        }
-        ++max;
-    }
-    return 0;
-}
